Wait for API initialization before probing music quiz state
useMusicQuizPlayer fetched quiz info on mount, but on a hard page
refresh the composable can mount before the API connection has
fetched server state. Probing that early reads an empty provider
registry and briefly resolves to a wrong 'no quiz' answer until a
provider event corrects it.

Promote the guest entry resolver's private wait-for-initialized
helper to a shared waitForApiInitialization in the api helpers and
gate the mount-time fetch on it.
